Custom text selection toolbar now works.
However there is one know bug: the toolbar appears in unexpected screen positions
(mostly at the very top) because we use LocalView.current to get a view which is then
used to invoke startActionMode() and most likely that's the global view which takes the whole
screen space (the root view where the compose is attached to). The default text selection
toolbar uses AndroidComposeView which apparently (I'm now sure here) is not the root view
but some kind of a wrapper.