Await for first down event without requiring it being unconsumed.
There is currently a weird behavior where this gesture detector is not receiving
any tap/untap gesture most likely because the down event is being consumed by some
scroll parent or whatever and this makes the "tap text anywhere to hide text selection"
not work. Hopefully, this will fix the issue.